backend/requirements.txt
asgiref==3.6.0
Django==4.1.6
flake8==6.0.0
mccabe==0.7.0
Pillow==9.4.0
pycodestyle==2.10.0
pyflakes==3.0.1
sqlparse==0.4.3
djangorestframework==3.14.0